fix(*): added the types folder
